requirements.txt
asgiref==3.7.2
backports.zoneinfo;python_version<"3.9"
bcrypt==4.0.1
Django==4.2.7
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
